Parser and Velocity Function
The file parser has yet to start working but the velocity function should work just fine.
I believe the parser correctly finds the files that the user inputs for, but exactness is needed with typing.
For the calculations done with total distance and velocity, I use the temp_dist variable both times which will make error handling easier if we desire to set a maximum distance travelled for each increment (using if statements of velocity otherwise).