File Parser
The File Parser is currently functioning but the errors with selecting the correct columns might occur again. More research into the read_csv function provided by Pandas should be done.
For this file, there should be derivative and other calculations done to the graph to provide more useful graphical analysis. Morevoer, we should try to make it possible to graph mutliple data types relative to time.