File Parser & Writing Update
The new method for saving files will be with csv files so that the file parser can easily read them, giving an advantage over the txt files.
As for the file parser, it now works with completely functioning labels for all the graphs. For multiple types of data, they are graphed on the same graph. For one type of data over another, the derivative of that plot is the subplot below it.