Real Time Graphing and Basic Velocity Functioning
The real time graphing works very well and will be upgraded in the future to use whatever data the user desires to input.
As for the velocity calculations, the stored values are correct but the (time - prev_time) part of the velocity function is always 0. This is a major problem, as it does not return an error but it does not work.
Moreover, the timestamp may not have been correctly updating at the actual rate. On personal observation, it felt a bit slow, but I could be wrong.

Tomorrow, velocity errors will be fixed, hopefully.