Velocity and Real Time Plot
The velocity function now works. The median method failed because the bin was too small and the medians of the current and previous would always be the same, which returned a velocity of 0.
Using the mean method, we get an error range of .1 m/s for the device standing still.
Real time plot has been implemented in the program to graph both x and y velocity.
The program could be more robot with more user input and hopefully additions to the GUI.
As of right now, everything seems to work well.