Fixed Errors Plotting in Position and Began Implementation of such into Motion Data
The errors with 3D positioning were fixed, as not enough variables were being returned when the program would encounter an error.
As for the motion data, the process of implementing real time plotting has begun, but the file is significantly different and will take much more effort to realize real time plotting.